requirements-pipeline.txt
requests==2.32.3
snowflake-connector-python==3.11.0
